Simplifying ProcessRawToArd dependencies
